+++ src/front/js/layout.js
@@ -2,7 +2,7 @@ import React from "react";
 import { BrowserRouter, Route, Switch } from "react-router-dom";
 //import ScrollToTop from "./component/scrollToTop";
 
-import { Contratos } from "./pages/contratos";
+import { Contratos } from "./pages/pages_contratos/contratos";
 import { Login } from "./pages/pages_globales/login";
 import injectContext from "./store/appContext";
 
@@ -12,19 +12,19 @@ import { Footer } from "./component/footer";
 
 //Nuestras paginas importadas
 import { Crearcontrato } from "./pages/crear_contrato";
-import { Datos_contrato } from "./pages/datos_contrato";
+import { Datos_contrato } from "./pages/pages_contratos/datos_contrato";
 import { Ordenes } from "./pages/ordenes";
 import { Datos_orden } from "./pages/datos_orden";
 import { Cuatrocerocuatro } from "./pages/pages_globales/404";
 import { Contactenos } from "./pages/pages_globales/contactanos";
-import { RestablecerContrasena } from "./pages/restablecer_contrasena";
+import { RestablecerContrasena } from "./pages/pages_globales/restablecer_contrasena";
 import { CambiarContrasena } from "./pages/pages_globales/cambiar_contrasena";
 import { VerFotosOrden } from "./pages/ver_fotos_orden";
 import { Listausuarios } from "./pages/pages_usuarios/listausers";
 import { Datosuser } from "./pages/pages_usuarios/Datosuser";
 import { Editaruser } from "./pages/pages_usuarios/edicionuser";
 import { Crearordentrabajo } from "./pages/crearordentrabajo";
-import { Editarcontrato } from "./pages/editar_contrato";
+import { Editarcontrato } from "./pages/pages_contratos/editar_contrato";
 
 //Nuestros componentes importados
 //import { Listatrabajo } from "./component/listatrabajo";
@@ -116,6 +116,10 @@ const Layout = () => {
 						<Datos_contrato />
 					</Route>
 					<Route exact path="/ordenes">
+						<Menusup />
+						<Ordenes />
+					</Route>
+					<Route exact path="/ordenes/:id">
 						<Menusup />
 						<Ordenes />
 					</Route>
